tests: Add utilities and fetch_file()
Handles compatibility utils between py2 and py3.
For example: unittest.mock, urllib, etc.

Also provides helpful utils for tests.

fetch_file():
Fetch a file based on the given key. If the file is not found, it is
created appropriately by either generating it or downloading it from
elsewhere. Useful to provide an abstraction level between finding
and using a file in all the tests.