ImageFile: Add analyze_softwares()
Analyze Softwares is a helper function which goes through the
exifdata and checked which software has modified the file and left
it's trace in the exif data. The exif data is messy, and each
software has different places where it leaves it's trace like the
Software field, Description field, Comment field, etc. This checks
all of them and makes a curated list of softwares.