ImageFile: Add analyze_face_haarcascades()
Use OpenCV's haarcascade to find faces and facial features.
This is a flexible method, as haarcascades can be trained for
other features too.

Note that opencv is not a required dependency in this commit.
I.e. It is skipped if cv2 is not importable. Also, a log message
is given mentioning that opencv is required for this analysis
routine.