ImageFile: Add more data to analyze_color_info()
The analyze_color_info() routine can now eben give EdgeRatio,
Number of greyscale shades used in the image, Percent of the color
pallette which are frequently used and so on. These are meant to
help as features to identify whether an image is a line drawing or
not. Hence, tests to compare line_drawings with non_line_Drawings
have also been added.